refactor: entity, peer, component as separate modules
