+++ packages/tsconfig/base.json
@@ -3,6 +3,7 @@
     "esModuleInterop": true,
     "skipLibCheck": true,
     "target": "es2022",
+    "lib": ["es2022", "dom"],
     "allowJs": true,
     "resolveJsonModule": true,
     "moduleDetection": "force",
